+++ cmd/server/main.go
@@ -13,6 +13,8 @@ import (
 	"log"
 	"os"
 
+	"github.com/CharVstack/CharV-backend/interfaces/middleware"
+
 	"github.com/CharVstack/CharV-backend/api"
 	"github.com/CharVstack/CharV-backend/infrastructure/disk"
 	"github.com/CharVstack/CharV-backend/infrastructure/file"
@@ -20,7 +22,6 @@ import (
 	"github.com/CharVstack/CharV-backend/infrastructure/utils"
 	"github.com/CharVstack/CharV-backend/interfaces/controller"
 	"github.com/CharVstack/CharV-backend/interfaces/router"
-	"github.com/CharVstack/CharV-backend/middleware"
 	"github.com/CharVstack/CharV-backend/usecase/host"
 	"github.com/CharVstack/CharV-backend/usecase/vm/qemu"
 	oapiMiddleware "github.com/deepmap/oapi-codegen/pkg/gin-middleware"
