+++ infrastructure/file/vm.go
@@ -92,6 +92,6 @@ func (v vmDataAccess) Add(vm entity.Vm) error {
 }
 
 func (v vmDataAccess) Delete(id uuid.UUID) error {
-	err := os.Remove(filepath.Join(v.path.Guests, id.String()+".json"))
-	return err
+	println(filepath.Join(v.path.Guests, id.String()+".json"))
+	return os.Remove(filepath.Join(v.path.Guests, id.String()+".json"))
 }
