+++ interfaces/router/router.go
@@ -4,7 +4,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/CharVstack/CharV-backend/middleware"
+	"github.com/CharVstack/CharV-backend/interfaces/middleware"
+
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
