+++ usecase/models/disk.go
@@ -4,6 +4,6 @@ import "github.com/google/uuid"
 
 // Disk of VM
 type Disk interface {
-	Create(id uuid.UUID) error
-	Delete(id uuid.UUID) error
+	Create(pool string, id uuid.UUID) error
+	Delete(pool string, id uuid.UUID) error
 }
